requirements.txt
flask
flask_sqlalchemy
torch
torchvision
pillow
werkzeug